Lower case "multiple/unknown language(s)" and "no linguistic content"
I counted the files in the CldrNames directory with these two regular
expressions:

'(mul|und|zxx)' => '\p{Ll}
'(mul|und|zxx)' => '\p{Lu}

Result: 141 of the strings for these three non-language codes start
with a lower case letter. 52 start with an upper case letter.

The strings are used in different contexts. In some of them it's
nice to have them all start with an upper case letter for *aestetic*
purposes, but this should be done with CSS (and must, because most
do not start with an upper case letter, see the counts above).

Discussion:
https://www.wikidata.org/w/index.php?title=Wikidata:Contact_the_development_team&oldid=359869978#Varying_caps_in_English_labels_for_language_codes_code_mul.2Czzx.2Cukn

Change-Id: Ia7a13540c39420a279b90d33d2317c87c1761263