Add LocalNames for Estonian
Languages that are not in CLDR (or use different code), but Wikipedia language version exists.

For the most part translations are taken from language name database by the Institute of the Estonian Language (http://www.eki.ee/knab/linguae.htm). A couple of names (missing from the database) are derived from other names.

Change-Id: I00943d3dba64f8a9cb0645722175f33e3f59354d