Rewrite confusing loop in CLDRParser
This loop was not looping anything, because the <identity> element
does not contain a list, but (as the name suggests) only singular,
identifying elements. This allows for quite some simplification.

Example:

<identity>
    <version number="$Revision$"/>
    <language type="en"/>
    <territory type="CA"/>
</identity>

Note that the fallback for the <language> element is not really needed,
because this element is guaranteed to be present. At least it is
present in all files at the moment. Since we can't guarantee this is
always the case, I think a fallback to the filename (without the .xml
file extension) is nice to have.

This change is motivated by Ifcf5b4a.

Change-Id: Ib0ddafa0b8113b82cf15953bd813c980d35266bb