Move the default entry on generated data to top
This is done by use of ! for the default keys.
The default keys are converted to const to make all the usages easier to
follow.

With the default entry at the top it is simpler to see a default value
unequal to the currency code and understand why all the languages needs
value different than the default.

Remove unneeded local names for the root language

Change-Id: Ifed7f0a9ec5f3ef6e934a901f486c68b92ea17c9