Remove Eastern Yiddish (ydd)
It's the same as Yiddish, for which the code "yi"
is always used.

It was added to translatewiki configuration for unknown
reasons before 2010, and it was never actually used for
anything.

I also made patches to remove it from translatewiki
( I3437c434eb02519797e758c9baeda236cd166cdb )
and from language-data
( https://github.com/wikimedia/language-data/pull/220 )

Change-Id: I3a30e3a1f199aae6d7e4acc6afb5fa7f13039312