Add LocalNamesNo.php and LocalNamesNn.php
Add Norwegian names for all languages the ULS knows about
that aren't already in CLDR data. Two exceptions: als and nrm,
both of which are codes which are used incorrectly in Wikimedia.

I've also added a few codes for some Sámi languages that aren't
already covered and some minority languages of Norway.

The reason I've used the language code `No` for the file name
and not `Nb` is to stay consistent with the CLDR data (which
uses `no` upstream for Norwegian Bokmål, and has that as a
fallback language for Norwegian Nynorsk; the corresponding
file in the CldrNames directory is CldrNamesNo.php.

Taking advantage of the fallback mechanisms, the Nynorsk file
only has definitions for language names which are different from
the names in Bokmål.

Change-Id: I31f48d22b894f410dfab0a36ca6ecb6169ba0dcb