Load CLDR localization for 'be-tarask'
The CLDR data file, at least in the v44 zip file, has TARASK in all
caps, but the proper language code is all lowercase. Add a hack to map
one to the other, enabling us to include previously missed be-tarask
CLDR data.

If you run the command

    git diff --no-index CldrNames/CldrNamesBe_tarask.php LocalNames/LocalNamesBe_tarask.php

you can see that there are all sorts of combinations between the two
sources of language names: many names are the same in both files (e.g.
'en' => 'ангельская'); many are only found in LocalNamesBe_tarask (e.g.
'hif' => 'фіджыйская гіндзі'), though often this is just because the
name is redundant with the base language and thus excluded from CLDR’s
be-tarask file (e.g. 'de' => 'нямецкая' is the same in CldrNamesBe);
many are only found in CldrNamesBe_tarask (e.g. 'bla' => 'блэкфуцкая');
and some also differ between the two sources (e.g. 'ca' => 'каталянская'
in CldrNamesBe_tarask but 'ca' => 'каталёнская' in LocalNamesBe_tarask –
я vs. ё as the third vowel).

Bug: T231755
Change-Id: I2572d08ed7486ece44b248b52a2d7a5231581a9a