Fix script for mni language
In CLDR, 'mni' means Meetei in Bengali script and 'mni-mtei' is Meetei
in Meetei Mayek; in MediaWiki, 'mni' means Meetei in Meetei Mayek, and
'mni-beng' would be Meetei in Bengali script (still in progress). Fix
rebuild.php (and rerun it) to move the CLDR data to Mni_beng.php,
leaving just a stub for Mni.php (will be extended via LocalNames in
Ieab5358b03).

Bug: T313883
Change-Id: Ib99f64e9c439e8fc7f10236b01a8f9db9aef31bd