Add additional LocalNamesEn
As we prepare to make the language names translatable on
translatewiki.net, we want to have English names (either via CldrNames
or LocalNames) for all languages that have a name in some other
languages, because otherwise there’s no source message for translators
to translate from; however, there are currently plenty of language codes
without an English name (mainly from LocalNamesNo, LocalNamesSh,
LocalNamesNds and LocalNamesFrr). This commit adds entries for them to
LocalNamesEn, as well as a test to ensure this doesn’t happen again in
future. The English names are mostly based on English Wikipedia
(following their handy [[ISO 639:$code]] redirects), with occasional
reference to other sources such as SIL’s website ([1] etc.) and the
LoC’s website [2].

For language codes that seem to refer to specific languages, I’ve
removed the “language” (or, in one case, “dialect”) suffix from the
English Wikipedia title. For language codes that seem to refer to
language families, I’ve kept the “languages” suffix. The three cases of
cau, nai and sai appear to not actually be language families, but merely
geographic groupings, but I’ve included them anyway; the names of nai
and sai attempt to strike a balance between the standard’s “North/South
American Indian languages” and English Wikipedia’s “Indigenous languages
of North/South America”. All of this is very much subject to change
depending on people who know this stuff better than I do, of course.

[1]: https://iso639-3.sil.org/code/afa
[1]: https://www.loc.gov/standards/iso639-2/php/English_list.php

Bug: T231755
Change-Id: Ia34e14ae0a1794b6123af6cc805b716a946b35dd
Depends-On: I315ac65ad0e6916a8fafdc7890443374e1870a1e
Depends-On: Ic3143a946b57d1ec35fbc0b205c6a65a248fd93d