Remove ruq-grek
This language was added long ago to translatewiki
and consequently to core MediaWiki.

However, it was never actually used for anything,
it was removed from core MediaWikia while ago,
and as bug T357981 says, it is never written in
this script in real life.

So there is no reason to keep entries for it
in our CLDR copy.

Most likely, ruq-cyrl will also be removed,
and ruq-latn will be merged with ruq, as
bug T357981 suggests, but I prefer to do these
things separately.

I have no prejudice against adding this variant
back if anyone produces proof that it is actually
used in real life.

Bug: T357981
Change-Id: Ie10ab0979a0d80ee1ab599b1a15201a185017b6d