Add a local Indonesian name for btd
This language is called "Batak Dairi" in English-language sources,
but "Batak Pakpak" in Indonesian-language sources.

Having this local name here will ensure that it's findable
in ULS by both names.

The reason it came up is that it was one of the languages
that were added to ULS (jquery.ime) keyboards lately
This is a part of cleaning up the code for those keyboards.

Bug: T377796
Change-Id: Ic578dd4c1d042b3501ef71cd3566290b1fe46b83