Change mak entry in LocalNamesEn to mak-bugi
CldrMain has had a mak entry (Makasar) since the initial commit in this
repository; change I2c09d40956 (commit 71741a535e) presumably meant to
add this name for mak-bugi, not mak, to make sure that mak-bugi also has
a language name in English (previously it was only in LocalNamesNo).
This also matches language-data [1], where mak without a script subtag
is Latin, not Buginese.

[1]: https://github.com/wikimedia/language-data/blob/a8ca74c446/data/langdb.yaml#L505

Bug: T231755
Change-Id: Ia898a498bde165c4469cad3f014306eb01f9f9f0